requirements.txt
Django==5.0.2
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
nltk==3.8.1
Pillow==10.1.0 